Fix issues. Seperate RadioGroup (in DialogFragmentRadio) from DialogFragment
