+++ app/src/main/java/gr/teicm/informatics/selfdrivegps/Activities/MapsActivity.java
@@ -13,6 +13,7 @@ import android.util.Log;
 import android.view.View;
 import android.widget.CompoundButton;
 import android.widget.ImageButton;
+import android.widget.ProgressBar;
 import android.widget.TextView;
 import android.widget.Toast;
 import android.widget.ToggleButton;
@@ -64,6 +65,7 @@ public class MapsActivity extends FragmentActivity
         ToggleButton mainStartBtn = findViewById(R.id.start_calculations); //Initialize view to make it invisible accordingly to mode
         ImageButton imageButtonForChangeMapTerrain = findViewById(R.id.bt_map_terrain_change);
         ImageButton imageButtonForChangeRangeMeter = findViewById(R.id.bt_change_range_meter);
+
         labelAboveToggleBtn = findViewById(R.id.tv_label_for_toggle_button); //Initialize view to change it accordingly to mode
         mSpeed = findViewById(R.id.tv_speed_of_user); //Initialize view for MapsUtilities.getSpecsForStatusBar
         mAccuracy = findViewById(R.id.tv_accuracy_of_gps); //Initialize view for MapsUtilities.getSpecsForStatusBar
@@ -73,14 +75,14 @@ public class MapsActivity extends FragmentActivity
 
         MapsUtilities.checkIfModeChanged(labelAboveToggleBtn, mainStartBtn);
 
+        //Call the DialogFragmentRadio /layout to set terrain on map
         imageButtonForChangeMapTerrain.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                controller.setLastProgramStatus(controller.getProgramStatus());
-                controller.setProgramStatus(Controller.MODE_0_SET_TERRAIN);
-                MapsUtilities.showAlertDialog(getFragmentManager());//Set listener on button to transfer data to database
+                MapsUtilities.showAlertDialogRadio(getFragmentManager());//Set listener on button to transfer data to database
             }
         });
+        //Call the DialogFragment /layout to set FieldName and RangeMeter
         imageButtonForChangeRangeMeter.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
